fix: ignored directories counts as a worktree directories
The plugin won't checkout a repository if we're currently in an ignored
directory of another repository. For example, let assume there's a
/opt/homebrew/ repository which has a var/buildkite-agent subdirectory,
ignored by the repo. If we cd into var/buildkite-agent, and run
git rev-parse --is-inside-work-tree, it will erroneously print "true". As a
result, the checkout plugin will operate on the /opt/homebrew repo and will
fail on the git clean step. To fix this, we need to ensure that the cwd is
not ignored by the repo.